FIX: Bizarre numbers shown in the tooltip for the maximum speed for way objects FIX: Certain simuconf.tab settings (cornering settings and instant path refresh) did not work if they were in the base simuconf.tab, rather than the pakset simuconf.tab FIX: Could "copy convoy" beyond credit limit. FIX: Interest not accounted for properly. CHANGE: Updated credits. CHANGE: Reduced slightly effect of poorer acceleration for steam locomotives.

+++ besch/way_obj_besch.h
@@ -80,9 +80,9 @@ private:
 	uint8 way_constraints_prohibitive;
 
 public:
-	long get_preis() const { return price; }
+	uint32 get_preis() const { return price; }
 
-	long get_wartung() const { return maintenance; }
+	uint32 get_wartung() const { return maintenance; }
 
 	/**
 	 * Determines max speed in km/h allowed on this way
